+++ app/models/produto.py
@@ -1,6 +1,4 @@
-from decimal import Decimal
-
-from sqlalchemy import Boolean, Integer, Numeric, String
+from sqlalchemy import Boolean, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base, TimestampMixin
@@ -13,6 +11,7 @@ class Produto(Base, TimestampMixin):
     sku: Mapped[str] = mapped_column(String(64), unique=True, index=True, nullable=False)
     nome: Mapped[str] = mapped_column(String(255), nullable=False)
     descricao: Mapped[str | None] = mapped_column(String(1000), nullable=True)
-    preco: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
+    # Preço em centavos (int). Dinheiro nunca é float nesta stack.
+    preco: Mapped[int] = mapped_column(Integer, nullable=False)
     quantidade_estoque: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
     ativo: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
