+++ app/schemas/produto.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from pydantic import BaseModel, ConfigDict, Field
 
 
@@ -7,14 +5,15 @@ class ProdutoCreate(BaseModel):
     sku: str = Field(min_length=1, max_length=64)
     nome: str = Field(min_length=1, max_length=255)
     descricao: str | None = Field(default=None, max_length=1000)
-    preco: Decimal = Field(gt=0, max_digits=12, decimal_places=2)
+    # Preço em centavos (int). Ex.: R$ 199,90 -> 19990.
+    preco: int = Field(gt=0)
     quantidade_estoque: int = Field(default=0, ge=0)
 
 
 class ProdutoUpdate(BaseModel):
     nome: str | None = Field(default=None, min_length=1, max_length=255)
     descricao: str | None = Field(default=None, max_length=1000)
-    preco: Decimal | None = Field(default=None, gt=0, max_digits=12, decimal_places=2)
+    preco: int | None = Field(default=None, gt=0)
     quantidade_estoque: int | None = Field(default=None, ge=0)
     ativo: bool | None = None
 
@@ -26,6 +25,6 @@ class ProdutoRead(BaseModel):
     sku: str
     nome: str
     descricao: str | None
-    preco: Decimal
+    preco: int
     quantidade_estoque: int
     ativo: bool
